app/requirements.txt
fastapi
uvicorn[standard]
websockets
openai
jinja2
python-multipart
python-dotenv
